ArrayAllocators

Documentation for ArrayAllocators.

This Julia package provides mechanisms to allocate arrays beyond that provided in the Base module of Julia.

The instances of the sub types of AbstractArrayAllocators take the place of undef in the Array{T}(undef, dims) invocation. This allows us to take advantage of alternative ways of allocating memory. The allocators take advantage of Base.unsafe_wrap in order to create arrays from pointers. A finalizer is also added for allocators that do not use Libc.free.

The original inspiration for this package is the memory allocator calloc. calloc allocates the memory and guarantees that the memory will be initialized by zeros. By this definition, it would appear equivalent to Base.zeros. However, calloc is potentially able to take advantage of operating system facilities that allocate memory lazily on demand rather than eagerly. Additonally, it may be able to obtain memory from the operating system that has already been initialized by zeros due to security constraints. On many systems, this allocator returns as quickly as malloc, the allocator used by Array{T}(undef, dims). In particular, in Python, numpy.zeros uses calloc, which may at times appear faster than Base.zeros in Julia.

In contrast, Base.zeros allocates memory using malloc and then uses fill! to eagerly and explicitly fill the array with zeros. On some systems, this may be a redudnant operation since the operating system may already know the allocated memory is filled with zeros.

This package makes calloc and other allocators available. Some of these allocators are specific to particular kinds of systems.

One example is allocating on Non-Uniform Memory Access (NUMA) nodes. On a NUMA system, random-access memory (RAM) may be accessible by certain processor cores at lower latency and higher bandwidth than other cores. Thus, it makes sense to allocate memory on particular NUMA nodes. On Linux, this is facilitated by numactl software which includes libnuma. On Windows, NUMA-aware memory allocation is exposed via the Kernel32 memory application programming interface such as via the function VirtualAllocExNuma. This package provides an abstraction over the two libraries.

Another application is memory alignment which may facilitate the use of advanced vector instructions in modern processors.

One other feature of this package is implementation of "Safe" allocators. These allocators provide extra protection by detecting integer overflow situations. Integer overflow can occur when multiplying large numbers causing the result to potentially wrap around. Memory allocators may report success after allocating an wrapped around number of bytes. The "Safe" allocators use the SaferIntegers to detect integer overflow avoiding this erroneous situation.

ArrayAllocators.calloc

ArrayAllocators.malloc

ArrayAllocators.safe_calloc

ArrayAllocators.safe_malloc

ArrayAllocators.AbstractArrayAllocator

ArrayAllocators.CallocAllocator

ArrayAllocators.MallocAllocator

ArrayAllocators.SafeCallocAllocator

ArrayAllocators.SaferMallocAllocator

ArrayAllocators.wrap_libc_pointer

ArrayAllocators.calloc — Constant

calloc

CallocAllocator singleton

ArrayAllocators.malloc — Constant

malloc

MallocAllocator singleton.

ArrayAllocators.safe_calloc — Constant

safe_calloc

SafeCallocAllocator singleton

ArrayAllocators.safe_malloc — Constant

safe_malloc

SafeMallocAllocator singleton

ArrayAllocators.AbstractArrayAllocator — Type

AbstractArrayAllocator

Parent abstract type for array allocators. Defines Array{T}(allocator, dims...) where T = Array{T}(allocator, dims)

ArrayAllocators.CallocAllocator — Type

CallocAllocator()

Use Libc.calloc to allocate an array. This is similar to zeros, except that the Libc implementation or the operating system may allocate and zero the memory in a lazy fashion.

ArrayAllocators.MallocAllocator — Type

MallocAllocator()

Allocate array using Libc.malloc. This is not meant to be useful but rather just to prototype the concept for a custom array allocator concept. This should be similar to using undef.

ArrayAllocators.SafeCallocAllocator — Type

SafeCallocAllocator()

Use SaferIntegers.SafeInt to calculate the number of bytes needed. See CallocAllocator.

ArrayAllocators.SaferMallocAllocator — Type

SafeMallocAllocator()

Use SaferInteger.SafeInt to calculate the number of bytes needed. See SaferMallocAllocator.

ArrayAllocators.wrap_libc_pointer — Method

wrap_libc_pointer(::Type{A}, ptr::Ptr{T}, dims) where {T, A <: AbstractArray{T}}
wrap_libc_pointer(ptr::Ptr{T}, dims) where {T, A <: AbstractArray{T}}

Checks to see if ptr is C_NULL for an OutOfMemoryError. Owns the array such that Libc.free is used.