ArrayAllocators

Documentation for ArrayAllocators.

Introduction

This Julia package provides mechanisms to allocate arrays beyond that provided in the Base module of Julia.

The instances of the sub types of AbstractArrayAllocator take the place of undef in the Array{T}(undef, dims) invocation. This allows us to take advantage of alternative ways of allocating memory. The allocators take advantage of Base.unsafe_wrap in order to create arrays from pointers. A finalizer is also added for allocators that do not use Libc.free.

In the base ArrayAllocators package, the following allocators are provided.

calloc

malloc

MemAlign(alignment)

An extension for use with Non-Uniform Memory Access allocations is available via the subpackage NumaAllocators.jl.

Example Basic Usage

Each of the methods below allocate 1 MiB of memory. Using undef as the first argument allocate uninitialized memory. The values are not guaranteed to be 0 or any other value.

In Base, the method zeros can be used to explicitly fill the memory with zeros. This is equivalent to using fill!(..., 0). Using calloc guarantees the values will be 0, yet is often as fast as using undef initialization.

julia> using ArrayAllocators

julia> @time U = Array{Int8}(undef, 1024, 1024);
  0.000019 seconds (2 allocations: 1.000 MiB)

julia> @time Z1 = zeros(Int8, 1024, 1024);
  0.000327 seconds (2 allocations: 1.000 MiB)

julia> @time Z2 = fill!(Array{UInt8}(undef, 1024, 1024), 0);
  0.000301 seconds (2 allocations: 1.000 MiB)

julia> @time C = Array{Int8}(calloc, 1024, 1024);
  0.000020 seconds (4 allocations: 1.000 MiB)

julia> sum(C)
0

Caveats

Above calloc appears to be much faster than zeros at generating an array full of 0s. However, some of the array created with zeros has already been fully allocated. The array allocated with calloc take longer to initialize since the operating system may have deferred the actual allocation of memory.

julia> @time Z = zeros(Int8, 1024, 1024);
  0.000324 seconds (2 allocations: 1.000 MiB)

julia> @time fill!(Z, 1);
  0.000138 seconds

julia> @time fill!(Z, 2);
  0.000136 seconds

julia> @time U = Array{Int8}(calloc, 1024, 1024);
  0.000020 seconds (4 allocations: 1.000 MiB)

julia> @time fill!(U, 1);
  0.000349 seconds

julia> @time fill!(U, 2);
  0.000136 seconds

Allocators

ArrayAllocators.calloc — Constant

calloc

CallocAllocator singleton instance. calloc will allocate memory and guarantee initialization to 0. See the type for details and the C standard library function for further details.

Example

julia> A = Array{UInt8}(calloc, 16, 16);

julia> sum(A)
0x0000000000000000

ArrayAllocators.malloc — Constant

malloc

MallocAllocator singleton instance. malloc will only allocate memory. It does not initialize memory is is similar in use as undef. See the type and the C standard library function for details.

Example

julia> Array{UInt8}(malloc, 16, 16);

Aligned Memory

ArrayAllocators.MemAlign — Type

MemAlign([alignment::Integer])

Allocate aligned memory. Alias for platform specific implementations.

alignment must be a power of 2.

On POSIX systems, alignment must be a multiple of sizeof(Ptr). On Windows, alignment must be a multiple of 2^16.

If alignment is not specified, it will be set to min_alignment(MemAlign).

MemAlign is a constant alias for one the following platform specific implementations.

POSIX (Linux and macOS): POSIX.PosixMemAlign

Windows: Windows.WinMemAlign.

ArrayAllocators.alignment — Function

alignment(alloc::AbstractMemAlign)

Get byte alignment of the AbstractMemAlign array allocator.

ArrayAllocators.min_alignment — Function

min_alignment(::AbstractMemAlign)

Get the minimum byte alignment of the AbstractMemAlign array allocator.

Types

ArrayAllocators.CallocAllocator — Type

CallocAllocator()

Use Libc.calloc to allocate an array. This is similar to zeros, except that the Libc implementation or the operating system may allocate and zero the memory in a lazy fashion.

See also https://en.cppreference.com/w/c/memory/calloc .

ArrayAllocators.MallocAllocator — Type

MallocAllocator()

Allocate array using Libc.malloc. This is not meant to be useful but rather just to prototype the concept for a custom array allocator concept. This should be similar to using undef.

See also https://en.cppreference.com/w/c/memory/malloc .

ArrayAllocators.UndefAllocator — Type

UndefAllocator{B}

Allocate arrays using the builtin undef method. The B parameter is a ByteCalculator

Internals

ArrayAllocators.AbstractArrayAllocator — Type

AbstractArrayAllocator{B}

Parent abstract type for array allocators. Parameter B is an AbstractByteCalculator Defines Array{T}(allocator, dims...) where T = Array{T}(allocator, dims)

ArrayAllocators.AbstractMemAlign — Type

AbstractMemAlign{B} <: AbstractArrayAllocator{B}

Abstract supertype for aligned memory allocators.

ArrayAllocators.wrap_libc_pointer — Function

wrap_libc_pointer(::Type{A}, ptr::Ptr{T}, dims) where {T, A <: AbstractArray{T}}
wrap_libc_pointer(ptr::Ptr{T}, dims) where {T, A <: AbstractArray{T}}

Checks to see if ptr is C_NULL for an OutOfMemoryError. Owns the array such that Libc.free is used.

Platform Specific Interface

Windows

ArrayAllocators.Windows.WinMemAlign — Type

WinMemAlign([alignment, lowestStartingAddress, highestStartingAddress])

Uses VirtualAlloc2 to allocate aligned memory. alignment must be a power of 2 and larger than 65536.

POSIX

ArrayAllocators.POSIX.PosixMemAlign — Type

PosixMemAlign([alignment::Integer])

Uses posix_memalign to allocate aligned memory. alignment must be a power of 2 and larger than sizeof(Ptr).

If alignment is provided, it will be set to min_alignment(PosixMemAlign).

Example

julia> Array{UInt8}(PosixMemAlign(32), 16, 16);